backend/requirements.txt
# Django Core
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Authentication
PyJWT==2.8.0

# Authentication
PyJWT==2.8.0
pyotp==2.9.0
qrcode==7.4.2

# Image Processing
Pillow==10.2.0

# PDF Generation (for reports)
reportlab==4.0.9

# Utilities
python-dotenv==1.0.1

# Development
black==24.1.1
pytest==8.0.0
pytest-django==4.7.0

torch>=2.0.0
transformers>=4.30.0
opencv-python>=4.7.0
numpy>=1.24.0
gdown>=4.7.0

grad-cam==1.5.0
opencv-python-headless>=4.5.0